+++ app/src/main/java/com/burrow/sensorActivity2/ui/captureHistory/CaptureHistoryUIState.kt
@@ -1,6 +1,6 @@
 package com.burrow.sensorActivity2.ui.captureHistory
 
 data class CaptureHistoryUIState (
-    val batchId: Int = 0,
+    val batchId: Long = 0,
     val timestamp : String = "DD MM YY HH.MM.SS.###"
 )
